api/requirements.txt
Django==3.1.1
djangorestframework==3.12.1
django-phone-verify==2.0.0
django-oauth-toolkit==1.3.2
django-searchable-encrypted-fields==0.1.6